move _copyRecursive to load
